Add files via upload
I made a script that I implemented into a bash bunny attack that it calls to pull the system information from a laptop and save it into the /loot/ directory of it. Once I adjust the ejection process of when the script is complete/saved onto the BB it should eject the (so called USB) and exit the ps session, I will post the script I used in the other directory. it is very insightful since it pulls the following:
*System info
*Accounts
*WLAN Profiles and Passwords
This is a great script since its a driveby attack for the early stages of setting up a more advanced attack
(possibly a reverse shell script coming soon)